♻️ extract domain context types to a separate file
