+++ packages/rum-core/src/domain/assembly.spec.ts
@@ -1,7 +1,8 @@
 import { ErrorSource, ONE_MINUTE, RawError, RelativeTime, display } from '@datadog/browser-core'
 import { createRawRumEvent } from '../../test/fixtures'
 import { setup, TestSetupBuilder } from '../../test/specHelper'
-import { CommonContext, RawRumErrorEvent, RawRumEvent, RumEventDomainContext, RumEventType } from '../rawRumEvent.types'
+import { RumEventDomainContext } from '../domainContext.types'
+import { CommonContext, RawRumErrorEvent, RawRumEvent, RumEventType } from '../rawRumEvent.types'
 import { RumActionEvent, RumErrorEvent, RumEvent } from '../rumEvent.types'
 import { startRumAssembly } from './assembly'
 import { LifeCycle, LifeCycleEventType, RawRumEventCollectedData } from './lifeCycle'
