+++ packages/rum-core/src/rawRumEvent.types.ts
@@ -6,7 +6,6 @@ import {
   ResourceType,
   ServerDuration,
   TimeStamp,
-  Omit,
 } from '@datadog/browser-core'
 
 export enum RumEventType {
@@ -220,49 +219,3 @@ export interface CommonContext {
   context: Context
   hasReplay?: true
 }
-
-export type RumEventDomainContext<T extends RumEventType = any> = T extends RumEventType.VIEW
-  ? RumViewEventDomainContext
-  : T extends RumEventType.ACTION
-  ? RumActionEventDomainContext
-  : T extends RumEventType.RESOURCE
-  ? RumFetchResourceEventDomainContext | RumXhrResourceEventDomainContext | RumOtherResourceEventDomainContext
-  : T extends RumEventType.ERROR
-  ? RumErrorEventDomainContext
-  : T extends RumEventType.LONG_TASK
-  ? RumLongTaskEventDomainContext
-  : never
-
-export interface RumViewEventDomainContext {
-  location: Readonly<Location>
-}
-export interface RumActionEventDomainContext {
-  event?: Event
-}
-export interface RumFetchResourceEventDomainContext {
-  requestInit?: RequestInit
-  requestInput: RequestInfo
-  response?: Response
-  error?: Error
-  performanceEntry?: PerformanceEntryRepresentation
-}
-export interface RumXhrResourceEventDomainContext {
-  xhr: XMLHttpRequest
-  performanceEntry?: PerformanceEntryRepresentation
-}
-export interface RumOtherResourceEventDomainContext {
-  performanceEntry: PerformanceEntryRepresentation
-}
-export interface RumErrorEventDomainContext {
-  error: unknown
-}
-export interface RumLongTaskEventDomainContext {
-  performanceEntry: PerformanceEntryRepresentation
-}
-
-/**
- * Symbolizes the type of the value returned by performanceEntry.toJSON(). Can also be built
- * manually to represent other kind of performance entries (ex: initial_document) or polyfilled
- * based on `performance.timing`.
- */
-export type PerformanceEntryRepresentation = Omit<PerformanceEntry, 'toJSON'>
